changed structure a bit
moved rigid bodies and colliders into their own files.
